+++ common-tools/clas-jcsg/src/main/java/org/jlab/detector/geant4/v2/MPGD2/URWT/URWTGeant4Factory.java
@@ -1,7 +1,7 @@
-package org.jlab.detector.geant4.v2.MPGD.URWT;
+package org.jlab.detector.geant4.v2.MPGD2.URWT;
 
 import org.jlab.detector.calib.utils.DatabaseConstantProvider;
-import org.jlab.detector.geant4.v2.MPGD.trapezoid.MPGDTrapezoidGeant4Factory;
+import org.jlab.detector.geant4.v2.MPGD2.trapezoid.MPGDTrapezoidGeant4Factory;
 
 /**
  * Geant4 factory for the uRWell Tracker (URWT).
