+++ reconstruction/urwt/src/main/java/org/jlab/service/urwt/URWRStrip.java
@@ -6,7 +6,7 @@ import org.jlab.detector.banks.RawDataBank;
 import org.jlab.detector.base.DetectorDescriptor;
 import org.jlab.detector.base.DetectorType;
 import org.jlab.detector.calib.utils.ConstantsManager;
-import org.jlab.detector.geant4.v2.MPGD.URWT.URWTStripFactory;
+import org.jlab.detector.geant4.v2.MPGD2.URWT.URWTStripFactory;
 import org.jlab.geom.prim.Line3D;
 import org.jlab.io.base.DataEvent;
